+++ docs/src/App.tsx
@@ -42,6 +42,7 @@ const AccordionConstrainedDemoPage = page(() => import('./pages/AccordionConstra
 const TabsDemoPage          = page(() => import('./pages/TabsDemo'));
 const SliderDemoPage        = page(() => import('./pages/SliderDemo'));
 const ProgressDemoPage      = page(() => import('./pages/ProgressDemo'));
+const SkeletonDemoPage      = page(() => import('./pages/SkeletonDemo'));
 const SelectDemoPage        = page(() => import('./pages/SelectDemo'));
 const TablePlaygroundPage   = page(() => import('./pages/TableDemo'));
 const ListDemoPage          = page(() => import('./pages/ListDemoPage'));
@@ -121,6 +122,7 @@ export function App() {
         <Route path="/tabs-demo"       element={<TabsDemoPage />} />
         <Route path="/slider-demo"     element={<SliderDemoPage />} />
         <Route path="/progress-demo"   element={<ProgressDemoPage />} />
+        <Route path="/skeleton-demo"   element={<SkeletonDemoPage />} />
         <Route path="/select-demo"     element={<SelectDemoPage />} />
         <Route path="/table-demo"      element={<TablePlaygroundPage />} />
         <Route path="/list-demo"       element={<ListDemoPage />} />